+++ validator/db/filesystem/attester_protection.go
@@ -131,10 +131,9 @@ func (s *Store) SaveAttestationForPubKey(
 	att ethpb.IndexedAtt,
 ) error {
 	// If there is no attestation, return on error.
-	if att == nil || att.GetData() == nil || att.GetData().Source == nil || att.GetData().Target == nil {
+	if att == nil || att.IsNil() || att.GetData().Source == nil || att.GetData().Target == nil {
 		return errors.New("incoming attestation does not contain source and/or target epoch")
 	}
-
 	// Get validator slashing protection.
 	validatorSlashingProtection, err := s.validatorSlashingProtection(pubkey)
 	if err != nil {